ギャラリー（ノート）の取得方法を変更、ページネーションに対応
